floppy: drive real disks through FluxBridge, with nothing vendored
The C++ FloppyDriveBridge and its ABI wrapper are gone -- the vendored
tree, the FFI, and the compiler that built them. In their place is
CopperlineHQ/FluxBridge, a pure-Rust library reached as an ordinary
dependency, so a build that says it supports a physical drive needs
nothing installed and carries no C++.

The module keeps the shape the emulator already spoke to, so the floppy
controller, the launcher and the guide are unchanged in substance: a
bridge still hands over one finished revolution as packed MFM plus the
length it wrapped at, still never waits outside the stalling mode, and
still says whether that revolution closes on itself. What it can now say
is better than a guess from the configured mode -- FluxBridge reports
what it proved about each capture, so an index-less one shown to be a
whole AmigaDOS track is trusted for replay like an index-aligned one.

Writes become asynchronous: submitted to the library, laid down on its
thread as the disk turns, and reported afterwards rather than waited on.

FluxBridge is itself a Rust port of the runtime parts of Rob Smith's
FloppyDriveBridge and carries that provenance in its own NOTICE.
